import from master
import from master

+++ PuppyWorld/src/main/java/com/example/ad/service/PostCommentService.java
@@ -8,7 +8,7 @@ public interface PostCommentService {
 	
 	public void savePostComment(PostComment postComment);
 	public ArrayList<PostComment> findAllPostComments();
-	public PostComment findPostCommentById(Long Id);
-	public void deleteEventById(Long Id);
+	public PostComment findPostCommentById(int Id);
+	public void deleteEventById(int Id);
 
 }
